+++ src/data/mod.rs
@@ -1,11 +1,12 @@
 use std::collections::HashMap;
-use std::fs;
-use std::io::{Cursor, Read};
+use std::fs::read;
+use std::io::Cursor;
 use std::path::PathBuf;
 use std::sync::{Arc, RwLock};
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
+use std::time::Duration;
 
 use anyhow::Context;
+use rdb::read_from_file;
 use tokio::spawn;
 use tokio::sync::broadcast::{Receiver, Sender};
 use tokio::task::JoinHandle;
@@ -15,78 +16,9 @@ use crate::encoding::{empty_string, okay_string};
 use crate::errors::{wrong_type, wrong_type_str};
 use crate::request::{self, CommandExpiration, SetOverride};
 use crate::utils::current_unix_timestamp;
-use crate::{encoding, transmission, utils};
+use crate::{encoding, transmission};
 
-// https://rdb.fnordig.de/file_format.html
-#[derive(PartialEq, Debug)]
-enum OpCode {
-    Eof,
-    SelectDB,
-    ExpireTime,
-    ExpireTimeMS,
-    ResizeDb,
-    Aux,
-    Other(u8),
-}
-
-impl OpCode {
-    fn from_byte(byte: u8) -> Self {
-        match byte {
-            0xFF => OpCode::Eof,
-            0xFE => OpCode::SelectDB,
-            0xFD => OpCode::ExpireTime,
-            0xFC => OpCode::ExpireTimeMS,
-            0xFB => OpCode::ResizeDb,
-            0xFA => OpCode::Aux,
-            other => OpCode::Other(other),
-        }
-    }
-}
-
-#[allow(dead_code)]
-enum AuxField {
-    RedisVersion,
-    RedisBits,
-    CreationTime,
-    MemoryUsed,
-}
-
-#[allow(dead_code)]
-#[derive(PartialEq, Debug)]
-enum ValueType {
-    String = 0,
-    List = 1,
-    Set = 2,
-    SortedSet = 3,
-    Hash = 4,
-    Zipmap = 9,
-    Ziplist = 10,
-    Intset = 11,
-    SortedSetZiplist = 12,
-    HashmapZiplist = 13,
-    ListQuicklist = 14,
-}
-
-impl ValueType {
-    fn from_byte(byte: u8) -> Result<Self, anyhow::Error> {
-        let value_type = match byte {
-            0 => Self::String,
-            1 => Self::List,
-            2 => Self::Set,
-            3 => Self::SortedSet,
-            4 => Self::Hash,
-            9 => Self::Zipmap,
-            10 => Self::Ziplist,
-            11 => Self::Intset,
-            12 => Self::SortedSetZiplist,
-            13 => Self::HashmapZiplist,
-            14 => Self::ListQuicklist,
-            val => anyhow::bail!("Unrecognized value type: {}", val),
-        };
-
-        Ok(value_type)
-    }
-}
+mod rdb;
 
 pub struct Database(Arc<RwLock<HashMap<String, DatabaseItem>>>);
 
@@ -114,6 +46,16 @@ impl Database {
         };
 
         Ok(data)
+    }
+
+    pub fn from_config(path: PathBuf) -> Result<Self, anyhow::Error> {
+        let database = Database::new();
+        if !path.exists() {
+            return Ok(database);
+        }
+
+        let contents = read(path).context("Reading RDB file")?;
+        read_from_file(database, Cursor::new(contents))
     }
 
     pub fn get_type(&self, key: &str) -> Option<String> {
@@ -576,67 +518,6 @@ impl Database {
 
         Ok(keys)
     }
-
-    pub fn from_config(path: PathBuf) -> Result<Self, anyhow::Error> {
-        let database = Database::new();
-        if !path.exists() {
-            return Ok(database);
-        }
-
-        let contents = fs::read(path).context("Reading RDB file")?;
-        let mut cursor = Cursor::new(contents);
-
-        let mut magic_string: [u8; 5] = [0; 5];
-        cursor
-            .read_exact(&mut magic_string)
-            .context("Reading magic string")?;
-        let magic_string =
-            String::from_utf8(magic_string.to_vec()).context("Parsing magic string into utf8")?;
-        if &magic_string != "REDIS" {
-            anyhow::bail!(
-                "Expected REDIS magic string at beginning of RDB file, got {}",
-                magic_string
-            );
-        }
-
-        let mut version_number: [u8; 4] = [0; 4];
-        cursor
-            .read_exact(&mut version_number)
-            .context("Reading version number")?;
-        let version_number = String::from_utf8(version_number.to_vec())
-            .context("Parsing verison number into utf8")?;
-        str::parse::<usize>(&version_number)
-            .context("Version number cannot be parsed as an integer")?;
-
-        loop {
-            let op_code = utils::read_next_byte(&mut cursor)?;
-            match OpCode::from_byte(op_code) {
-                OpCode::Aux => parse_aux(&mut cursor)?,
-                OpCode::SelectDB => parse_select_db(&mut cursor)?,
-                OpCode::ResizeDb => parse_resize_db(&mut cursor)?,
-                OpCode::ExpireTimeMS => {
-                    let database_item = parse_expire_time_ms(&mut cursor)?;
-                    if let Some((key, value)) = database_item {
-                        database.set_item(key, value);
-                    }
-                }
-                OpCode::ExpireTime => {
-                    let database_item = parse_expire_time_sec(&mut cursor)?;
-                    if let Some((key, value)) = database_item {
-                        database.set_item(key, value);
-                    }
-                }
-                OpCode::Other(value_type_byte) => {
-                    let value_type = ValueType::from_byte(value_type_byte)?;
-                    let (key, value) = read_key_value_pair(value_type, None, &mut cursor)?;
-                    database.set_item(key, value);
-                }
-                OpCode::Eof => break,
-            }
-        }
-
-        Ok(database)
-    }
 }
 
 impl Clone for Database {
@@ -739,111 +620,6 @@ pub struct ReadStreamItem<'a> {
 pub struct TempReadStreamItem {
     pub streams: Vec<InnerRedisStream>,
     pub key: String,
-}
-
-#[derive(Debug)]
-pub struct AuxValue {
-    #[allow(dead_code)]
-    key: String,
-    #[allow(dead_code)]
-    value: String,
-}
-
-fn parse_aux(cursor: &mut Cursor<Vec<u8>>) -> Result<(), anyhow::Error> {
-    let key = encoding::decode_rdb_string(cursor)?;
-    let value = encoding::decode_rdb_string(cursor)?;
-
-    let _aux_value = AuxValue { key, value };
-
-    Ok(())
-}
-
-fn parse_select_db(cursor: &mut Cursor<Vec<u8>>) -> Result<(), anyhow::Error> {
-    let _size = encoding::decode_rdb_int(cursor)?;
-
-    Ok(())
-}
-
-fn parse_resize_db(cursor: &mut Cursor<Vec<u8>>) -> Result<(), anyhow::Error> {
-    let _hash_table_size = encoding::decode_rdb_int(cursor)?;
-    let _expiry_table_size = encoding::decode_rdb_int(cursor)?;
-
-    Ok(())
-}
-
-fn parse_expire_time_ms(
-    cursor: &mut Cursor<Vec<u8>>,
-) -> Result<Option<(String, DatabaseItem)>, anyhow::Error> {
-    let mut expire_time_ms: [u8; 8] = [0; 8];
-    cursor.read_exact(&mut expire_time_ms)?;
-    let expire_time_milliseconds = u64::from_le_bytes(expire_time_ms);
-    let expire_time_unix_timestamp = expire_time_milliseconds / 1000;
-
-    read_expirable_item(expire_time_unix_timestamp, cursor)
-}
-
-fn parse_expire_time_sec(
-    cursor: &mut Cursor<Vec<u8>>,
-) -> Result<Option<(String, DatabaseItem)>, anyhow::Error> {
-    let mut expire_time_seconds: [u8; 4] = [0; 4];
-    cursor.read_exact(&mut expire_time_seconds)?;
-    let expire_time_seconds = u32::from_le_bytes(expire_time_seconds);
-
-    read_expirable_item(expire_time_seconds as u64, cursor)
-}
-
-fn read_expirable_item(
-    expire_time_unix_timestamp: u64,
-    cursor: &mut Cursor<Vec<u8>>,
-) -> Result<Option<(String, DatabaseItem)>, anyhow::Error> {
-    let item_expiration = duration_to_item_expiration(expire_time_unix_timestamp);
-    let item_expires_in_future = item_expiration.is_some();
-
-    let value_type_byte = utils::read_next_byte(cursor)?;
-    let value_type = ValueType::from_byte(value_type_byte)?;
-
-    let item_data = read_key_value_pair(value_type, item_expiration, cursor)?;
-
-    if item_expires_in_future {
-        let data = Some(item_data);
-        Ok(data)
-    } else {
-        Ok(None)
-    }
-}
-
-fn read_key_value_pair(
-    value_type: ValueType,
-    expire_time: Option<Duration>,
-    cursor: &mut Cursor<Vec<u8>>,
-) -> Result<(String, DatabaseItem), anyhow::Error> {
-    let key = encoding::decode_rdb_string(cursor)?;
-    let value = match value_type {
-        ValueType::String => encoding::decode_rdb_string(cursor)?,
-        // TODO
-        _ => anyhow::bail!("{:?} value type not supported", value_type),
-    };
-
-    let redis_string = RedisString::new(value, expire_time);
-    let database_item = DatabaseItem::String(redis_string);
-
-    Ok((key, database_item))
-}
-
-fn duration_to_item_expiration(expire_time_unix_timestamp: u64) -> Option<Duration> {
-    let now = SystemTime::now();
-    let duration_since_epoch = now.duration_since(UNIX_EPOCH).unwrap();
-
-    // Convert the duration to seconds and return it as u64
-    let current_unicode_timestamp = duration_since_epoch.as_secs();
-
-    match expire_time_unix_timestamp.checked_sub(current_unicode_timestamp) {
-        Some(dur) => {
-            let duration = Duration::from_secs(dur);
-            Some(duration)
-        }
-        None => None,
-    }
 }
 
 fn determine_sequence_number(
